+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 [//]: # (The format is based on [Keep a Changelog]&#40;https://keepachangelog.com/en/1.1.0/&#41;, and this project adheres to [Semantic Versioning]&#40;https://semver.org/spec/v2.0.0.html&#41;.)
 
+## [0.4.4] - dev
+
+### Changed
+
+- Fix shader patch error in `ClearcoatTintPlugin` material extension
+- Add `AssetImporter.autoSetName` (default `true`), to control whether the asset name should be set automatically from the file name when importing from files/blobs.
+- Add `TransformControlsPlugin.selectionFilterTest` to filter/swap objects that can be selected for transform controls.
+- Set `allowOverride` `false` for materials in `TransformControls`
+
 ## [0.4.3] - 2025-12-01
 
 ### Changed